+++ examples/regional_private_node_pool_oauth_scopes/main.tf
@@ -16,7 +16,7 @@
 
 module "gke" {
   source  = "terraform-google-modules/kubernetes-engine/google//modules/private-cluster"
-  version = "~> 40.0"
+  version = "~> 43.0"
 
   project_id                        = var.project_id
   name                              = "random-test-cluster"