+++ examples/simple_regional_with_networking/main.tf
@@ -53,7 +53,7 @@ module "gcp-network" {
 
 module "gke" {
   source  = "terraform-google-modules/kubernetes-engine/google"
-  version = "~> 40.0"
+  version = "~> 43.0"
 
   project_id             = var.project_id
   name                   = var.cluster_name