+++ modules/gke-standard-cluster/variables.tf
@@ -34,7 +34,6 @@ variable "project_id" {
 variable "location" {
   description = "The location (region or zone) in which the cluster master will be created, as well as the default node location. If you specify a zone (such as us-central1-a), the cluster will be a zonal cluster with a single cluster master. If you specify a region (such as us-west1), the cluster will be a regional cluster with multiple masters spread across zones in the region, and with default node locations in those zones as well."
   type        = string
-  default     = null
 }
 
 variable "network" {
@@ -309,7 +308,7 @@ variable "master_auth" {
 }
 
 variable "master_authorized_networks_config" {
-  description = "The desired configuration options for master authorized networks."
+  description = "The desired configuration options for master authorized networks. Cidr Block must follow [Cidr notation](https://en.wikipedia.org/wiki/Classless_Inter-Domain_Routing#CIDR_notation)"
   type = object({
     cidr_blocks = list(object({
       display_name = string