+++ test/fixtures/simple_windows_node_pool/outputs.tf
@@ -42,7 +42,8 @@ output "master_kubernetes_version" {
 }
 
 output "kubernetes_endpoint" {
-  value = module.this.kubernetes_endpoint
+  value     = module.this.kubernetes_endpoint
+  sensitive = true
 }
 
 output "client_token" {